% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{pf_custom}
\alias{pf_custom}
\title{Calculate F Distribution Cumulative Probability}
\usage{
pf_custom(F, df1, df2)
}
\arguments{
\item{F}{The F-statistic value.}

\item{df1}{Degrees of freedom for the numerator.}

\item{df2}{Degrees of freedom for the denominator.}
}
\value{
The cumulative probability associated with the F-statistic.
}
\description{
This helper function computes the cumulative distribution function (CDF) for the F-distribution.
It is used to calculate the p-value for the F-statistic in model performance evaluation.
}
