% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_coeff.R
\name{simple_linear_regression_coeffs}
\alias{simple_linear_regression_coeffs}
\title{simple_linear_regression_coeffs}
\usage{
simple_linear_regression_coeffs(user_dat, x, y)
}
\arguments{
\item{user_dat}{user data}

\item{x}{choice of covariate}

\item{y}{choice of response variable}
}
\value{
The estimated intercept and slope of the simple linear regression model
}
\description{
Calculate the estimated covariate' coefficients for simple linear regression
}
\examples{
df1 <- data.frame(x = c(1, 2, 3, 4, 5), y = c(1, 5, 8, 15, 26))
simple_linear_regression_coeffs(df1, x = df1$x, y = df1$y)

# Source of data: https://www.datacamp.com/tutorial/linear-regression-R
library(readxl)
ageandheight <- read_excel("ageandheight.xls", sheet = "Hoja2")
simple_linear_regression_coeffs(ageandheight, x = ageandheight$age,
                                y = ageandheight$height)

}
