% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_coeff.R
\name{simple_linear_regression_coeffs}
\alias{simple_linear_regression_coeffs}
\title{simple_linear_regression_coeffs}
\usage{
simple_linear_regression_coeffs(user_dat, x, y)
}
\arguments{
\item{user_dat}{user data}

\item{x}{choice of covariate}

\item{y}{choice of response variable}
}
\value{
The estimated intercept and slope of the simple linear regression model
}
\description{
Calculate the estimated covariate' coefficients for simple linear regression
}
\examples{
df1 <- data.frame(x = c(1, 2, 3, 4, 5), y = c(1, 5, 8, 15, 26))
simple_linear_regression_coeffs(df1, x = df1$x, y = df1$y)

ctl <- c(4.17, 5.58, 5.18, 6.11, 4.50, 4.61, 5.17, 4.53, 5.33, 5.14)
trt <- c(4.81, 4.17, 4.41, 3.59, 5.87, 3.83, 6.03, 4.89, 4.32, 4.69)
group <- gl(2, 10, 20, labels = c("0", "1"))
weight <- c(ctl, trt)
df2 <- data.frame(weight, group)
simple_linear_regression_coeffs(df2, x = as.numeric(group), y = weight)

}
